mines_telegram_bot/requirements.txt
aiogram==3.0.0b7
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
python-dotenv==1.0.0
cryptography==41.0.3
